Add fuctionality to transfer to level 2
Add class GoToLevel2 to appear a window interface with the score which Hero has gained.

Add a counter in Counter class to count the enemies that the Hero has killed to achieve the goal of the level.

Add a method to Hero class to checking if achieve the goal of the level.

Add a getter to Level_1_World for the counter of enemies that Hero kills.

Add parameters to the contractor of Level_2_World.